Create qis subpackage and move some existing code to qis.states and qis.measures (#2808)
Moves the following things to `linalg/states.py` from `sim/wave_function.py` and `sim/density_matrix_utils.py`:
```
STATE_VECTOR_LIKE
bloch_vector_from_state_vector
density_matrix_from_state_vector
dirac_notation
to_valid_state_vector
validate_normalized_state
to_valid_density_matrix
von_neumann_entropy
```
This may not be their final resting place (see https://github.com/quantumlib/Cirq/pull/2797#discussion_r387915534) but it certainly makes more sense than before, given that we have a file called `states.py`. Also, it has the benefit of removing some circular dependency import hacks.
