+++ cirq/linalg/states_test.py
@@ -12,49 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import numpy as np
-
 import cirq
+from cirq._compat_test import capture_logging
 
 
-def test_one_hot():
-    result = cirq.one_hot(shape=4, dtype=np.int32)
-    assert result.dtype == np.int32
-    np.testing.assert_array_equal(result, [1, 0, 0, 0])
+def test_deprecated():
+    with capture_logging() as log:
+        _ = cirq.linalg.eye_tensor((1,), dtype=float)
+    assert len(log) == 1
+    assert "cirq.eye_tensor" in log[0].getMessage()
+    assert "deprecated" in log[0].getMessage()
 
-    np.testing.assert_array_equal(
-        cirq.one_hot(shape=[2, 3], dtype=np.complex64), [[1, 0, 0], [0, 0, 0]])
-
-    np.testing.assert_array_equal(
-        cirq.one_hot(shape=[2, 3], dtype=np.complex64, index=(0, 2)),
-        [[0, 0, 1], [0, 0, 0]])
-
-    np.testing.assert_array_equal(
-        cirq.one_hot(shape=5, dtype=np.complex128, index=3), [0, 0, 0, 1, 0])
-
-
-def test_eye_tensor():
-    assert np.all(cirq.eye_tensor((), dtype=int) == np.array(1))
-    assert np.all(cirq.eye_tensor((1,), dtype=int) == np.array([[1]]))
-    assert np.all(cirq.eye_tensor((2,), dtype=int) == np.array([
-        [1, 0],
-        [0, 1]]))  # yapf: disable
-    assert np.all(cirq.eye_tensor((2, 2), dtype=int) == np.array([
-        [[[1, 0], [0, 0]],
-         [[0, 1], [0, 0]]],
-        [[[0, 0], [1, 0]],
-         [[0, 0], [0, 1]]]]))  # yapf: disable
-    assert np.all(cirq.eye_tensor((2, 3), dtype=int) == np.array([
-        [[[1, 0, 0], [0, 0, 0]],
-         [[0, 1, 0], [0, 0, 0]],
-         [[0, 0, 1], [0, 0, 0]]],
-        [[[0, 0, 0], [1, 0, 0]],
-         [[0, 0, 0], [0, 1, 0]],
-         [[0, 0, 0], [0, 0, 1]]]]))  # yapf: disable
-    assert np.all(cirq.eye_tensor((3, 2), dtype=int) == np.array([
-        [[[1, 0], [0, 0], [0, 0]],
-         [[0, 1], [0, 0], [0, 0]]],
-        [[[0, 0], [1, 0], [0, 0]],
-         [[0, 0], [0, 1], [0, 0]]],
-        [[[0, 0], [0, 0], [1, 0]],
-         [[0, 0], [0, 0], [0, 1]]]]))  # yapf: disable
+    with capture_logging() as log:
+        _ = cirq.linalg.one_hot(shape=(1,), dtype=float)
+    assert len(log) == 1
+    assert "cirq.one_hot" in log[0].getMessage()
+    assert "deprecated" in log[0].getMessage()
