+++ cirq/qis/__init__.py
@@ -1,0 +1,18 @@
+from cirq.qis.measures import (
+    fidelity,
+    von_neumann_entropy,
+)
+
+from cirq.qis.states import (
+    bloch_vector_from_state_vector,
+    density_matrix_from_state_vector,
+    dirac_notation,
+    eye_tensor,
+    one_hot,
+    STATE_VECTOR_LIKE,
+    to_valid_density_matrix,
+    to_valid_state_vector,
+    validate_indices,
+    validate_qid_shape,
+    validate_normalized_state,
+)
