+++ cirq/qis/measures.py
@@ -1,4 +1,4 @@
-# Copyright 2020 The Cirq Developers
+# Copyright 2019 The Cirq Developers
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,10 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Distance measures between quantum states and operations."""
+"""Measures on and between quantum states and operations."""
 
 import numpy as np
 import scipy
+import scipy.stats
 
 
 def _sqrt_positive_semidefinite_matrix(mat: np.ndarray) -> np.ndarray:
@@ -57,3 +58,14 @@ def fidelity(state1: np.ndarray, state2: np.ndarray) -> float:
         return trace**2
     raise ValueError('The given arrays must be one- or two-dimensional. '
                      f'Got shapes {state1.shape} and {state2.shape}.')
+
+
+def von_neumann_entropy(density_matrix: np.ndarray) -> float:
+    """Calculates von Neumann entropy of density matrix in bits.
+    Args:
+        density_matrix: The density matrix.
+    Returns:
+        The calculated von Neumann entropy.
+    """
+    eigenvalues = np.linalg.eigvalsh(density_matrix)
+    return scipy.stats.entropy(abs(eigenvalues), base=2)
