+++ cirq/sim/density_matrix_utils_test.py
@@ -18,214 +18,21 @@ import numpy as np
 import pytest
 
 import cirq
+from cirq._compat_test import capture_logging
 
 
-def assert_valid_density_matrix(matrix, num_qubits=None, qid_shape=None):
-    if qid_shape is None and num_qubits is None:
-        num_qubits = 1
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(matrix,
-                                     num_qubits=num_qubits,
-                                     qid_shape=qid_shape,
-                                     dtype=matrix.dtype), matrix)
+def test_deprecated():
+    with capture_logging() as log:
+        _ = cirq.sim.to_valid_density_matrix(0, 1)
+    assert len(log) == 1
+    assert "cirq.to_valid_density_matrix" in log[0].getMessage()
+    assert "deprecated" in log[0].getMessage()
 
-
-def test_von_neumann_entropy():
-    # 1x1 matrix
-    assert cirq.von_neumann_entropy(np.array([[1]])) == 0
-    # An EPR pair state (|00> + |11>)(<00| + <11|)
-    assert cirq.von_neumann_entropy(
-        np.array([1, 0, 0, 1] * np.array([[1], [0], [0], [1]]))) == 0
-    # Maximally mixed state
-    # yapf: disable
-    assert cirq.von_neumann_entropy(np.array(
-        [[0.5, 0],
-        [0, 0.5]])) == 1
-    # 3x3 state
-    assert np.isclose(cirq.von_neumann_entropy(
-        np.array(
-            [[0.5, 0.5j, 1],
-            [-0.5j, 0.5, 0],
-            [0.7, 0.4, 0.6]])),
-                      1.37,
-                      atol=1e-01)
-    # 4X4 state
-    assert np.isclose(cirq.von_neumann_entropy(
-        np.array(
-            [[0.5, 0.5j, 1, 3],
-            [-0.5j, 0.5, 0, 4],
-            [0.7, 0.4, 0.6, 5],
-            [6, 7, 8, 9]])),
-                      1.12,
-                      atol=1e-01)
-    # yapf: enable
-    # 2x2 random unitary, each column as a ket, each ket as a density matrix,
-    # linear combination of the two with coefficients 0.1 and 0.9
-    res = cirq.testing.random_unitary(2)
-    first_column = res[:, 0]
-    first_density_matrix = 0.1 * np.outer(first_column, np.conj(first_column))
-    second_column = res[:, 1]
-    second_density_matrix = 0.9 * np.outer(second_column,
-                                           np.conj(second_column))
-    assert np.isclose(cirq.von_neumann_entropy(first_density_matrix +
-                                               second_density_matrix),
-                      0.4689,
-                      atol=1e-04)
-
-    assert np.isclose(cirq.von_neumann_entropy(
-        np.diag([0, 0, 0.1, 0, 0.2, 0.3, 0.4, 0])),
-                      1.8464,
-                      atol=1e-04)
-
-
-def test_to_valid_density_matrix_from_density_matrix():
-    assert_valid_density_matrix(np.array([[1, 0], [0, 0]]))
-    assert_valid_density_matrix(np.array([[0.5, 0], [0, 0.5]]))
-    assert_valid_density_matrix(np.array([[0.5, 0.5], [0.5, 0.5]]))
-    assert_valid_density_matrix(np.array([[0.5, 0.2], [0.2, 0.5]]))
-    assert_valid_density_matrix(np.array([[0.5, 0.5j], [-0.5j, 0.5]]))
-    assert_valid_density_matrix(
-        np.array([[0.5, 0.2 - 0.2j], [0.2 + 0.2j, 0.5]]))
-    assert_valid_density_matrix(np.eye(4) / 4.0, num_qubits=2)
-    assert_valid_density_matrix(np.diag([1, 0, 0, 0]), num_qubits=2)
-    assert_valid_density_matrix(np.ones([4, 4]) / 4.0, num_qubits=2)
-    assert_valid_density_matrix(np.diag([0.2, 0.8, 0, 0]), num_qubits=2)
-    assert_valid_density_matrix(np.array(
-        [[0.2, 0, 0, 0.2 - 0.3j],
-         [0, 0, 0, 0],
-         [0, 0, 0, 0],
-         [0.2 + 0.3j, 0, 0, 0.8]]),
-        num_qubits=2)
-
-    assert_valid_density_matrix(np.array([[1, 0, 0]] + [[0, 0, 0]] * 2),
-                                qid_shape=(3,))
-    assert_valid_density_matrix(np.array([[0, 0, 0], [0, 0.5, 0.5j],
-                                          [0, -0.5j, 0.5]]),
-                                qid_shape=(3,))
-    assert_valid_density_matrix(np.eye(9) / 9.0, qid_shape=(3, 3))
-    assert_valid_density_matrix(np.eye(12) / 12.0, qid_shape=(3, 4))
-    assert_valid_density_matrix(np.ones([9, 9]) / 9.0, qid_shape=(3, 3))
-    assert_valid_density_matrix(np.diag([0.2, 0.8, 0, 0]), qid_shape=(4,))
-
-
-def test_to_valid_density_matrix_not_square():
-    with pytest.raises(ValueError, match='square'):
-        cirq.to_valid_density_matrix(np.array([[1, 0]]), num_qubits=1)
-    with pytest.raises(ValueError, match='square'):
-        cirq.to_valid_density_matrix(np.array([[1], [0]]), num_qubits=1)
-
-
-def test_to_valid_density_matrix_size_mismatch_num_qubits():
-    with pytest.raises(ValueError, match='size'):
-        cirq.to_valid_density_matrix(np.array([[1, 0], [0, 0]]), num_qubits=2)
-    with pytest.raises(ValueError, match='size'):
-        cirq.to_valid_density_matrix(np.eye(4) / 4.0, num_qubits=1)
-
-
-def test_to_valid_density_matrix_not_hermitian():
-    with pytest.raises(ValueError, match='hermitian'):
-        cirq.to_valid_density_matrix(np.array([[1, 0.1], [0, 0]]), num_qubits=1)
-    with pytest.raises(ValueError, match='hermitian'):
-        cirq.to_valid_density_matrix(np.array([[0.5, 0.5j], [0.5, 0.5j]]),
-                                     num_qubits=1)
-    with pytest.raises(ValueError, match='hermitian'):
-        cirq.to_valid_density_matrix(
-            np.array(
-                [[0.2, 0, 0, -0.2 - 0.3j],
-                 [0, 0, 0, 0],
-                 [0, 0, 0, 0],
-                 [0.2 + 0.3j, 0, 0, 0.8]]),
-            num_qubits=2)
-
-
-def test_to_valid_density_matrix_mismatched_qid_shape():
-    with pytest.raises(ValueError, match=r'num_qubits != len\(qid_shape\)'):
-        cirq.to_valid_density_matrix(np.eye(4) / 4,
-                                     num_qubits=1,
-                                     qid_shape=(2, 2))
-    with pytest.raises(ValueError, match=r'num_qubits != len\(qid_shape\)'):
-        cirq.to_valid_density_matrix(np.eye(4) / 4,
-                                     num_qubits=2,
-                                     qid_shape=(4,))
-    with pytest.raises(ValueError, match='Both were None'):
-        cirq.to_valid_density_matrix(np.eye(4) / 4)
-
-
-def test_to_valid_density_matrix_not_unit_trace():
-    with pytest.raises(ValueError, match='trace 1'):
-        cirq.to_valid_density_matrix(np.array([[1, 0], [0, 0.1]]), num_qubits=1)
-    with pytest.raises(ValueError, match='trace 1'):
-        cirq.to_valid_density_matrix(np.array([[1, 0], [0, -0.1]]),
-                                     num_qubits=1)
-    with pytest.raises(ValueError, match='trace 1'):
-        cirq.to_valid_density_matrix(np.zeros([2, 2]), num_qubits=1)
-
-
-def test_to_valid_density_matrix_not_positive_semidefinite():
-    with pytest.raises(ValueError, match='positive semidefinite'):
-        cirq.to_valid_density_matrix(
-            np.array([[1.1, 0], [0, -0.1]], dtype=np.complex64), num_qubits=1)
-    with pytest.raises(ValueError, match='positive semidefinite'):
-        cirq.to_valid_density_matrix(
-            np.array([[0.6, 0.5], [0.5, 0.4]], dtype=np.complex64),
-            num_qubits=1)
-
-
-def test_to_valid_density_matrix_wrong_dtype():
-    with pytest.raises(ValueError, match='dtype'):
-        cirq.to_valid_density_matrix(
-            np.array([[1, 0], [0, 0]], dtype=np.complex64),
-            num_qubits=1, dtype=np.complex128)
-
-
-def test_to_valid_density_matrix_from_state():
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(
-            density_matrix_rep=np.array([1, 0], dtype=np.complex64),
-            num_qubits=1),
-        np.array([[1, 0], [0, 0]]))
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(
-            density_matrix_rep=np.array([np.sqrt(0.3), np.sqrt(0.7)],
-                                        dtype=np.complex64),
-            num_qubits=1),
-        np.array([[0.3, np.sqrt(0.3 * 0.7)], [np.sqrt(0.3 * 0.7), 0.7]]))
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(
-            density_matrix_rep=np.array([np.sqrt(0.5), np.sqrt(0.5) * 1j],
-                                        dtype=np.complex64),
-            num_qubits=1),
-        np.array([[0.5, -0.5j], [0.5j, 0.5]]))
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(
-            density_matrix_rep=np.array([0.5] * 4, dtype=np.complex64),
-            num_qubits=2),
-        0.25 * np.ones((4, 4)))
-
-
-def test_to_valid_density_matrix_from_state_invalid_state():
-    with pytest.raises(ValueError, match="shape was neither"):
-        cirq.to_valid_density_matrix(np.array([1, 0, 0]), num_qubits=2)
-
-
-def test_to_valid_density_matrix_from_computational_basis():
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(density_matrix_rep=0, num_qubits=1),
-        np.array([[1, 0], [0, 0]]))
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(density_matrix_rep=1, num_qubits=1),
-        np.array([[0, 0], [0, 1]]))
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(density_matrix_rep=2, num_qubits=2),
-        np.array([[0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 1, 0], [0, 0, 0, 0]]))
-    np.testing.assert_almost_equal(
-        cirq.to_valid_density_matrix(density_matrix_rep=0, num_qubits=0),
-        np.array([[1]]))
-
-
-def test_to_valid_density_matrix_from_state_invalid_computational_basis():
-    with pytest.raises(ValueError, match="out of range"):
-        cirq.to_valid_density_matrix(-1, num_qubits=2)
+    with capture_logging() as log:
+        _ = cirq.sim.von_neumann_entropy(np.eye(2) / 2)
+    assert len(log) == 1
+    assert "cirq.von_neumann_entropy" in log[0].getMessage()
+    assert "deprecated" in log[0].getMessage()
 
 
 def test_sample_density_matrix_big_endian():
