+++ docs/api.rst
@@ -216,14 +216,10 @@ results.
 
     cirq.CIRCUIT_LIKE
     cirq.RANDOM_STATE_OR_SEED_LIKE
-    cirq.STATE_VECTOR_LIKE
     cirq.big_endian_bits_to_int
     cirq.big_endian_digits_to_int
     cirq.big_endian_int_to_bits
     cirq.big_endian_int_to_digits
-    cirq.bloch_vector_from_state_vector
-    cirq.density_matrix_from_state_vector
-    cirq.dirac_notation
     cirq.final_density_matrix
     cirq.final_wavefunction
     cirq.flatten
@@ -240,10 +236,7 @@ results.
     cirq.sample_sweep
     cirq.to_resolvers
     cirq.to_sweep
-    cirq.to_valid_density_matrix
-    cirq.to_valid_state_vector
     cirq.validate_mixture
-    cirq.validate_normalized_state
     cirq.validate_probability
     cirq.xeb_fidelity
     cirq.CircuitSampleJob
@@ -617,8 +610,6 @@ Algebra and Representation
     cirq.diagonalize_real_symmetric_matrix
     cirq.dot
     cirq.expand_matrix_in_orthogonal_basis
-    cirq.eye_tensor
-    cirq.fidelity
     cirq.hilbert_schmidt_inner_product
     cirq.is_diagonal
     cirq.is_hermitian
@@ -638,7 +629,6 @@ Algebra and Representation
     cirq.map_eigenvalues
     cirq.match_global_phase
     cirq.matrix_from_basis_coefficients
-    cirq.one_hot
     cirq.partial_trace
     cirq.reflection_matrix_pow
     cirq.slice_for_qubits_equal_to
@@ -647,12 +637,30 @@ Algebra and Representation
     cirq.targeted_conjugate_about
     cirq.targeted_left_multiply
     cirq.unitary_eig
-    cirq.von_neumann_entropy
     cirq.wavefunction_partial_trace_as_mixture
     cirq.AxisAngleDecomposition
     cirq.Duration
     cirq.KakDecomposition
     cirq.Timestamp
+
+
+Quantum Information Science
+'''''''''''''''''''''''''''
+
+.. autosummary::
+    :toctree: generated/
+
+    cirq.STATE_VECTOR_LIKE
+    cirq.bloch_vector_from_state_vector
+    cirq.density_matrix_from_state_vector
+    cirq.dirac_notation
+    cirq.eye_tensor
+    cirq.fidelity
+    cirq.one_hot
+    cirq.to_valid_density_matrix
+    cirq.to_valid_state_vector
+    cirq.validate_normalized_state
+    cirq.von_neumann_entropy
 
 
 Internal Implementation Details
